feat: add GetLocationIDByBrokerCode functionality and update price offer location IDs to use location reference

+++ backend/services/booking/db/query/price_offers_query.sql
@@ -49,15 +49,15 @@ RETURNING *;
 -- name: GetPriceOfferByToken :one
 SELECT price_offers.* , pl.name AS pickup_location, dl.name AS dropoff_location
 FROM price_offers
-    JOIN locations pl ON price_offers.pickup_location_id::bigint = pl.id
-    JOIN locations dl ON price_offers.dropoff_location_id::bigint = dl.id
+    JOIN locations pl ON price_offers.pickup_location_id = pl.id
+    JOIN locations dl ON price_offers.dropoff_location_id = dl.id
  WHERE token = sqlc.arg(token);
 
 -- name: GetPriceOfferById :one
 SELECT price_offers.* , pl.name AS pickup_location, dl.name AS dropoff_location
 FROM price_offers
-    JOIN locations pl ON price_offers.pickup_location_id::bigint = pl.id
-    JOIN locations dl ON price_offers.dropoff_location_id::bigint = dl.id
+    JOIN locations pl ON price_offers.pickup_location_id = pl.id
+    JOIN locations dl ON price_offers.dropoff_location_id = dl.id
 WHERE price_offers.id = sqlc.arg(id) AND price_offers.agent_id = sqlc.arg(agent_id);
 
 -- name: UpdatePriceOffer :exec
@@ -76,8 +76,8 @@ pickup_date, return_date, pickup_time, dropoff_time,
 currency_code, total_price, offered_currency_code, offered_price, 
 price_offers.created_at
 FROM price_offers
-    JOIN locations pl ON price_offers.pickup_location_id::bigint = pl.id
-    JOIN locations dl ON price_offers.dropoff_location_id::bigint = dl.id
+    JOIN locations pl ON price_offers.pickup_location_id = pl.id
+    JOIN locations dl ON price_offers.dropoff_location_id = dl.id
 WHERE agent_id = sqlc.arg(agent_id)
   AND (sqlc.narg(status)::offer_status IS NULL OR status = sqlc.narg(status)::offer_status)
   AND (sqlc.narg(name_search)::text IS NULL OR price_offers.name ILIKE '%' || sqlc.narg(name_search)::text || '%')
